refactor: reorganize functions into separate files and add types
