+++ src/constants/index.ts
@@ -1,0 +1,6 @@
+export const SESSION_ID_STORAGE_KEY = "jb_widget_session_id";
+export const SESSION_ID_TTL = 24 * 60 * 60 * 1000; 
+export const supImgFiles = ["png", "jpg", "jpeg", "gif", "bmp", "webp", "image"];
+export const ALLOWED_IMAGE_INPUT_EXTENSIONS = ["png", "jpg", "jpeg"];
+export const ALLOWED_IMAGE_MIME_TYPES = ["image/png", "image/jpg", "image/jpeg"];
+export const FILE_LIMIT =  1 * 1024 * 1024;
