+++ src/types/index.ts
@@ -1,4 +1,5 @@
 
+// Type definitions for journey-builder-chat
 export type MainChatWidgetProps = {
   api_key?: string;
   output_type: string,
@@ -48,6 +49,7 @@ export type MainChatWidgetProps = {
   allow_web_hook: boolean;
 }
 
+// Props for the ChatWindow component
 export type ChatWindowProps = {
   api_key?: string;
   output_type: string,
@@ -100,6 +102,7 @@ export type ChatWindowProps = {
   allow_web_hook: boolean;
 }
 
+// Type for the ChatMessage 
 export type ChatMessageType = {
   message: string;
   isSend: boolean;
@@ -118,6 +121,7 @@ export type ChatMessageType = {
   text?: string
 };
 
+// Props for the FilePreview component
 export type FilePreviewProps = {
   loading: boolean;
   file: File;
@@ -135,10 +139,12 @@ export type FilePreviewProps = {
   retry_send_file_btn_style?: React.CSSProperties;
 };
 
+// Props for the ChatMessagePlaceholder component
 export type ChatMessagePlaceholderType = {
   bot_message_style?: React.CSSProperties;
 };
 
+// Type for file with additional properties
 export type file = { 
   file:File, 
   file_path?: string, 
@@ -148,3 +154,19 @@ export type file = {
   type: string 
 };
 
+export type WebhookeMessage = {
+  message: string,
+  sender: string,
+  timestamp: string,
+}
+
+export type WebhookResponse = {
+  messages: WebhookeMessage[]
+}
+
+export type WebhookDataToSend = {
+  session_id: string,
+  message: string,
+  origen: string,
+  stream: boolean
+}
